Updating timer to v2

+++ LIB/STD_TYPES.h
@@ -4,12 +4,12 @@
 	// Decimal type unsigned value with 8 bits size
 	typedef unsigned char u8;
 	typedef signed char s8;
-	typedef unsigned int u16;
-	typedef signed int s16;
+	typedef unsigned short u16;
+	typedef signed short s16;
 	typedef unsigned long u32;
 	typedef signed long s32;
-	typedef unsigned long long u64;
-	typedef signed long long s64;
+	typedef unsigned long long int u64;
+	typedef signed long long int s64;
 	typedef float f32;
 	typedef double f64;
